+++ Release Notes.md
@@ -32,6 +32,7 @@ SOFTWARE.
 * arm_cmsis: removed including HAL config (already included from CMSIS includes)
 * arm_hal_peripheral: added to include used peripherals
 * arm_hal_peripheral: reworked all peripheral includes
+* arm_chip_stm32: added wrapper for PWM tim/channels
 * arm_chip_xx: Families divided into files with respective fam name
 * refactoring names of platform header files